Fast access to Contrib instance; restructure contribs (#277)
* Contrib instance in registry; generalize commands

* Faster access to ContribInstances; pass args to sub-command hooks; restructure contribs
