+++ ludwig/cli.py
@@ -65,27 +65,27 @@ Available sub-commands:
         getattr(self, args.command)()
 
     def experiment(self):
-        ludwig.contrib.contrib_command("experiment")
+        ludwig.contrib.contrib_command("experiment", *sys.argv)
         experiment.cli(sys.argv[2:])
 
     def train(self):
-        ludwig.contrib.contrib_command("train")
+        ludwig.contrib.contrib_command("train", *sys.argv)
         train.cli(sys.argv[2:])
 
     def predict(self):
-        ludwig.contrib.contrib_command("predict")
+        ludwig.contrib.contrib_command("predict", *sys.argv)
         predict.cli(sys.argv[2:])
 
     def visualize(self):
-        ludwig.contrib.contrib_command("visualize")
+        ludwig.contrib.contrib_command("visualize", *sys.argv)
         visualize.cli(sys.argv[2:])
 
     def collect_weights(self):
-        ludwig.contrib.contrib_command("collect_weights")
+        ludwig.contrib.contrib_command("collect_weights", *sys.argv)
         collect.cli_collect_weights(sys.argv[2:])
 
     def collect_activations(self):
-        ludwig.contrib.contrib_command("collect_activations")
+        ludwig.contrib.contrib_command("collect_activations", *sys.argv)
         collect.cli_collect_activations(sys.argv[2:])
