+++ ludwig/contrib.py
@@ -16,72 +16,12 @@
 
 """
 Module for handling contributed support.
-All classes must have the following functions:
-- import_call: Loaded first to handle items
-  that need to be setup before core modules,
-  like tensorflow.
-If a call doesn't apply, provide an empy
-implementation with `pass`.
 """
 
 import logging
 import sys
 
-
-## Contributors, add classes here:
-
-class Comet():
-    """
-    Class that defines the methods necessary
-    """
-    @staticmethod
-    def import_call(argv, *args, **kwargs):
-        """
-        Enable Third-party support from comet.ml
-        Allows experiment tracking, visualization, and
-        management.
-        """
-        try:
-            import comet_ml
-        except ImportError:
-            logging.error("Ignored --comet: Please install comet_ml")
-            return None
-
-        return Comet()
-
-    def experiment(self, *args, **kwargs):
-        import comet_ml
-        self.experiment = comet_ml.Experiment(log_code=False)
-        config = comet_ml.get_config()
-        self._save_config(config)
-
-    def train(self, *args, **kwargs):
-        import comet_ml
-        self.experiment = comet_ml.Experiment(log_code=False)
-        config = comet_ml.get_config()
-        self._save_config(config)
-
-    def visualize(self, *args, **kwargs):
-        import comet_ml
-        self.experiment = comet_ml.ExistingExperiment()
-
-    def predict(self, *args, **kwargs):
-        import comet_ml
-        self.experiment = comet_ml.ExistingExperiment()
-
-    def _save_config(self, config):
-        config["comet.experiment_key"] = self.experiment.id
-        config.save()
-
-
-## Contributors, add your class here:
-contrib_registry = {
-    'comet': {
-        "class": Comet,
-        "instance": None,
-        }
-}
-
+from .contribs import contrib_registry
 
 def contrib_import():
     """
@@ -94,16 +34,16 @@ def contrib_import():
     """
     argv_list = sys.argv
     argv_set = set(argv_list)
-    for contrib_name in contrib_registry:
+    for contrib_name in contrib_registry["classes"]:
         parameter_name = '--' + contrib_name
         if parameter_name in argv_set:
             ## Calls ContribClass.import_call(argv_list)
             ## and return an instance, if appropriate
-            contrib_class = contrib_registry[contrib_name]["class"]
+            contrib_class = contrib_registry["classes"][contrib_name]
             instance = contrib_class.import_call(argv_list)
             ## Save instance in registry
             if instance:
-                contrib_registry[contrib_name]["instance"] = instance
+                contrib_registry["instances"].append(instance)
             ## Clean up and remove your flag
             sys.argv.remove(parameter_name)
 
@@ -114,9 +54,7 @@ def contrib_command(command, *args, **kwargs):
 
     ContribInstance.COMMAND(*args, **kwargs)
     """
-    for contrib_name in contrib_registry:
-        instance = contrib_registry[contrib_name]["instance"]
-        if instance:
-            method = getattr(instance, command, None)
-            if method:
-                method(*args, **kwargs)
+    for instance in contrib_registry["instances"]:
+        method = getattr(instance, command, None)
+        if method:
+            method(*args, **kwargs)
